services/graphql-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
strawberry-graphql==0.216.1
python-multipart==0.0.6
asyncpg==0.29.0
sqlalchemy==2.0.23
redis==5.0.1